moved debug menu classes to DebugMenu
